mcp-server/requirements.txt
fastmcp==0.5.0
requests==2.31.0
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.5.3
uvicorn==0.24.0
cryptography==41.0.8
aiofiles==23.2.1
